Modernize
Allow caller to customize logging
Update package references
Update publishing workflow
