+++ .github/workflows/publish.yml
@@ -1,0 +1,32 @@
+name: Publish
+
+on:
+  release:
+    types: [ published ]
+
+jobs:
+  publish:
+    if: startsWith(github.repository, 'chris-peterson')
+
+    runs-on: ubuntu-latest
+
+    env:
+      Configuration: Release
+      PackageReleaseNotes: ${{ github.event.release.body }}
+      PackageReleaseTag: ${{ github.event.release.tag_name }}
+
+    steps:
+      - uses: actions/checkout@v4
+      - name: Setup .NET
+        uses: actions/setup-dotnet@v4
+        with:
+          dotnet-version: 8.0.x
+
+      - name: Build
+        run: dotnet build
+      - name: Publish to nuget.org
+        run: dotnet nuget push **/*.nupkg --skip-duplicate --api-key ${{ secrets.NUGET_API_KEY }} --source https://api.nuget.org/v3/index.json
+
+    environment:
+      name: nuget.org
+      url: https://nuget.org/packages/Assurance
