+++ src/assurance/Assurance.csproj
@@ -1,26 +1,24 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <Version>1.1.3</Version>
+    <Version>1.2.0</Version>
     <TargetFramework>netstandard2.1</TargetFramework>
+    <LangVersion>latest</LangVersion>
     <Title>Assurance</Title>
     <Authors>Chris Peterson</Authors>
     <Description>A library to boost confidence when making code changes</Description>
-    <Copyright>2021-2023</Copyright>
+    <Copyright>2021-2025</Copyright>
     <AssemblyName>Assurance</AssemblyName>
     <PackageId>Assurance</PackageId>
     <PackageProjectUrl>https://github.com/chris-peterson/assurance</PackageProjectUrl>
     <PackageTags>validator</PackageTags>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageReleaseNotes>
-Include release notes in package
-    </PackageReleaseNotes>
     <PackageReadmeFile>README.md</PackageReadmeFile>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Spiffy.Monitoring" Version="6.0.*" />
+    <PackageReference Include="Spiffy.Monitoring" Version="6.4.*" />
   </ItemGroup>
 
   <ItemGroup>
