+++ tests/UnitTests/RunnerTests/_Contexts.cs
@@ -1,18 +1,20 @@
 using System;
 using System.Threading;
+using System.Threading.Tasks;
 using Spiffy.Monitoring;
 
-namespace Assurance.UnitTests
+namespace Assurance.UnitTests.RunnerTests
 {
-    public class RunnerTestContext
+    public class TestingContext
     {
         public Func<string> Existing { get; set; }
         public Func<string> Replacement { get; set; }
         public RunResult<string> Result { get; set; }
         public LogEvent LoggedEvent { get; set; }
+        
         public ManualResetEvent WaitForLogEvent = new ManualResetEvent(false);
 
-        public RunnerTestContext()
+        public TestingContext()
         {
             Configuration.Initialize(c =>
                 c.Providers.Add("custom", evt => {
@@ -21,4 +23,12 @@ namespace Assurance.UnitTests
                 }));
         }
     }
+    
+    public class AsyncTestingContext
+    {
+        public Func<Task<string>> Existing { get; set; }
+        public Func<Task<string>> Replacement { get; set; }
+        public RunResult<string> Result { get; set; }
+    }
+
 }
